// src/interfaces/response.ts
// Generated by https://quicktype.io

export interface ClubResponse {
  count: number;
  count_total: number;
  page: number;
  page_total: number;
  items_per_page: number;
  items: ItemClub[];
}

export interface ClubIDResponse {
  item: ItemClub;
}

export interface ItemClub {
  id: number;
  name: string;
  league: number;
}

export interface PlayersReponse {
  count: number;
  count_total: number;
  page: number;
  page_total: number;
  items_per_page: number;
  items: ItemPlayer[];
}

export interface ItemPlayer {
  id: number;
  resource_id: null;
  name: string;
  age: number;
  resource_base_id: null;
  fut_bin_id: null;
  fut_wiz_id: null;
  first_name: null | string;
  last_name: null | string;
  common_name: string;
  height: number;
  weight: number;
  birth_date: string;
  league: number;
  nation: number;
  club: number;
  rarity: number;
  traits: any[];
  specialities: any[];
  position: string;
  skill_moves: number;
  weak_foot: number;
  foot: Foot;
  attack_work_rate: WorkRate;
  defense_work_rate: WorkRate;
  total_stats: null;
  total_stats_in_game: null;
  rating: number;
  rating_average: null;
  pace: number;
  shooting: number;
  passing: number;
  dribbling: number;
  defending: number;
  physicality: number;
  pace_attributes: null;
  shooting_attributes: null;
  passing_attributes: null;
  dribbling_attributes: null;
  defending_attributes: null;
  physicality_attributes: null;
  goalkeeper_attributes: null;
}

export enum WorkRate {
  High = "High",
  Low = "Low",
  Med = "Med",
}

export enum Foot {
  Left = "Left",
  Right = "Right",
}

export interface Player {
  name: string;
  nation: string;
  club: string;
  position: string;
}

export interface NationResponse {
  item: ItemNation;
}

export interface ItemNation {
  id: number;
  name: string;
}

export interface NationGetResponse {
  count: number;
  count_total: number;
  page: number;
  page_total: number;
  items_per_page: number;
  items: ItemNationGetResponse[];
}

export interface ItemNationGetResponse {
  id: number;
  name: string;
}

export interface TeamResponse {
  page: number;
  totalPages: string;
  items: number;
  totalItems: number;
  players: Player[];
}